newsltr/requirements.txt
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
black==23.9.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.5
defusedxml==0.8.0rc2
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.6
django-anymail==10.2
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-health-check==3.17.0
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==6.0.1
django-trix==0.3.0
django-unfold==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djet==0.3.0
djoser==2.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
filelock==3.12.4
flake8==6.1.0
gunicorn==21.2.0
identify==2.5.29
idna==3.4
inflection==0.5.1
install==1.3.5
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.4.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2==2.9.9
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==7.1.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8
